requirements.txt
amqp==2.6.1
atomicwrites==1.4.1
attrs==21.4.0
backports.functools-lru-cache==1.6.4
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
click==7.1.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==5.0.1
ConfigArgParse==0.15.1
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==3.3.2
Django==1.11.29
django-tastypie==0.13.3
enum34==1.1.10
fancycompleter==0.9.1
Flask==1.1.4
funcsigs==1.0.2
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.16
humanfriendly==10.0
idna==2.10
importlib-metadata==2.1.3
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==4.6.11
locustio==0.13.5
MarkupSafe==1.1.1
monotonic==1.6
more-itertools==5.0.0
msgpack-python==0.5.6
packaging==20.9
pathlib2==2.3.7.post1
pdbpp==0.10.3
pluggy==0.13.1
py==1.11.0
pycparser==2.21
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.7
pyreadline==2.1
pyrepl==0.9.0
pytest==4.6.11
pytest-django==3.10.0
python-dateutil==2.8.2
python-dotenv==0.18.0
python-mimeparse==1.6.0
pytz==2022.6
pyzmq==19.0.2
requests==2.27.1
scandir==1.10.0
six==1.16.0
structlog==20.1.0
typing==3.10.0.0
urllib3==1.26.13
vine==1.3.0
wcwidth==0.2.5
Werkzeug==1.0.1
wmctrl==0.4
zipp==1.2.0
zope.event==4.5.0
zope.interface==5.5.2
